Add files via upload
Using the Flckr API to pull 2 pictures - for now. 
Right now, everything is hard coded. I had issues getting data back and getting the data to display. I was able to work through that, and I think I coded things to make it 'easy' to add variables for adding the second API and user input.